+++ src/Opulence/Sessions/Ids/Generators/IdGenerator.php
@@ -8,8 +8,6 @@
  */
 namespace Opulence\Sessions\Ids\Generators;
 
-use Opulence\Cryptography\Utilities\Strings;
-
 /**
  * Defines the session Id generator
  */
@@ -17,23 +15,21 @@ class IdGenerator implements IIdGenerator
 {
     /** The default length of an Id */
     const DEFAULT_LENGTH = 40;
-    /** @var Strings The strings utility */
-    private $strings = null;
-
-    /**
-     * @param Strings $strings The strings utility
-     */
-    public function __construct(Strings $strings)
-    {
-        $this->strings = $strings;
-    }
 
     /**
      * @inheritdoc
      */
     public function generate($length = self::DEFAULT_LENGTH)
     {
-        return $this->strings->generateRandomString($length);
+        // N bytes becomes 2N characters in bin2hex(), hence the division by 2
+        $string = bin2hex(random_bytes(ceil($length / 2)));
+
+        if ($length % 2 == 1) {
+            // Slice off one character to make it the appropriate odd length
+            $string = mb_substr($string, 1);
+        }
+
+        return $string;
     }
 
     /**